requirements.txt
certifi==2018.4.16
Django==1.11
django-js-asset==1.1.0
django-mptt==0.9.0
django-telegrambot==1.0.1
future==0.16.0
Pillow==5.2.0
python-telegram-bot==10.1.0
pytz==2018.5
